requirements.txt
Django==1.11
django-elasticsearch-dsl==0.5.0
django-elasticsearch-dsl-drf==0.16.2
django-filter==2.0.0
django-nine==0.1.13
django-rest-auth==0.9.3
djangorestframework==3.8.2
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
freeze==1.0.10
ipaddress==1.0.22
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
urllib3==1.24
